+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Fixed
+* Fixed an issue where the "Remove Bone Influences" button in the Mesh Editor did not actually remove influences.
+
 ## 0.10.4 - 2024-07-23
 ### Added
 * Added an option to remove unused entries in the Modl Editor.