+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added an option to convert a parent bone to bone influences in the Mesh Editor.
 * Added anisotropic rotation to the shader information in the Matl Editor.
 
+### Changed
+* Improved layout consistency for editors using a grid layout.
+
 ### Fixed
 * Fixed an issue where the "Remove Bone Influences" button in the Mesh Editor did not actually remove influences.
 * Fixed an issue where some materials would render with the wrong anisotropic highlight rotation.